# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/movies/search": {
      "get": {
        "operationId": "MoviesController_searchMovies",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "description": "Search term",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultsDto"
                }
              }
            }
          }
        },
        "summary": "Search movies from TMDB",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/discover": {
      "get": {
        "operationId": "MoviesController_discoverMovies",
        "parameters": [
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort order for results",
            "schema": {
              "default": "popularity.desc",
              "type": "string",
              "enum": [
                "popularity.desc",
                "popularity.asc",
                "release_date.desc",
                "release_date.asc",
                "vote_average.desc",
                "vote_average.asc"
              ]
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "description": "Filter by release year",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultsDto"
                }
              }
            }
          }
        },
        "summary": "Discover popular movies from TMDB",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/tmdb/{movieId}": {
      "get": {
        "operationId": "MoviesController_getMovieDetails",
        "parameters": [
          {
            "name": "movieId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TMDBMovieDetailDto"
                }
              }
            }
          }
        },
        "summary": "Get movie details from TMDB",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/tmdb/{movieId}/credits": {
      "get": {
        "operationId": "MoviesController_getFullMovieCredits",
        "parameters": [
          {
            "name": "movieId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullCreditsDto"
                }
              }
            }
          }
        },
        "summary": "Get the full cast and crew for a movie, crew grouped by department",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/tmdb/{movieId}/watch-providers": {
      "get": {
        "operationId": "MoviesController_getWatchProviders",
        "parameters": [
          {
            "name": "movieId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "description": "ISO 3166-1 country code (e.g. US, GB)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchProvidersResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get watch providers for a movie from TMDB/JustWatch",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/tmdb/{movieId}/recommendations": {
      "get": {
        "operationId": "MoviesController_getRecommendations",
        "parameters": [
          {
            "name": "movieId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultsDto"
                }
              }
            }
          }
        },
        "summary": "Get TMDB recommendations (similar movies)",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/user/{userDid}": {
      "get": {
        "operationId": "MoviesController_getUserMovies",
        "parameters": [
          {
            "name": "userDid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrackedMovieDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get tracked movies for a user",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/watched": {
      "post": {
        "operationId": "MoviesController_markWatched",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "movieId"
                ],
                "properties": {
                  "movieId": {
                    "type": "string",
                    "description": "TMDB movie ID"
                  },
                  "watchedAt": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true,
                    "description": "Custom watch datetime (ISO 8601). Null creates an undated Watch; omission uses the current time."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackedMovieDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "Mark a movie as watched",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/watched/{movieId}": {
      "delete": {
        "operationId": "MoviesController_unmarkWatched",
        "parameters": [
          {
            "name": "movieId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "description": "Remove mode: latest (default) removes most recent watch, all removes all watches",
            "schema": {
              "enum": [
                "latest",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Movie unmarked as watched"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "Unmark a movie as watched",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/{movieId}": {
      "get": {
        "operationId": "MoviesController_getMovie",
        "parameters": [
          {
            "name": "movieId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDto"
                }
              }
            }
          }
        },
        "summary": "Get movie from database",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/user/{userDid}/movie/{movieId}/history": {
      "get": {
        "operationId": "MoviesController_getMovieWatchHistory",
        "parameters": [
          {
            "name": "userDid",
            "required": true,
            "in": "path",
            "description": "User DID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "movieId",
            "required": true,
            "in": "path",
            "description": "TMDB movie ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Watch history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WatchHistoryItemDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "summary": "Get watch history for a specific movie",
        "tags": [
          "movies"
        ]
      }
    },
    "/movies/history/{trackedMovieId}": {
      "delete": {
        "operationId": "MoviesController_deleteWatchHistoryEntry",
        "parameters": [
          {
            "name": "trackedMovieId",
            "required": true,
            "in": "path",
            "description": "Tracked movie entry ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Watch history entry deleted"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Tracked movie entry not found"
          }
        },
        "summary": "Delete a specific watch history entry",
        "tags": [
          "movies"
        ]
      }
    },
...
    "/search/discover": {
      "get": {
        "operationId": "SearchController_discoverAll",
        "parameters": [
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort by",
            "schema": {
              "default": "popularity.desc",
              "type": "string",
              "enum": [
                "popularity.desc",
                "popularity.asc",
                "vote_average.desc",
                "vote_average.asc",
                "release_date.desc",
                "release_date.asc",
                "primary_release_date.desc",
                "primary_release_date.asc"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "description": "Year filter",
            "schema": {
              "minimum": 1900,
              "maximum": 2100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedDiscoverResponseDto"
                }
              }
            }
          }
        },
        "summary": "Discover popular movies and shows from TMDB",
        "tags": [
          "search"
        ]
      }
    },
...
  },
  "info": {
    "title": "Opnshelf API",
    "description": "Personal media tracker powered by AT Protocol",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
      "TMDBMovieResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "poster_path": {
            "type": "string"
          },
          "backdrop_path": {
            "type": "string"
          },
          "release_date": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "vote_average": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "title"
        ]
      },
      "SearchResultsDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total count of items across all pages"
          },
          "page": {
            "type": "number",
            "description": "Current page number after server-side clamping"
          },
          "pageSize": {
            "type": "number",
            "description": "Number of items per page"
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of available pages"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Whether a next page exists"
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "Whether a previous page exists"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TMDBMovieResultDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage",
          "items"
        ]
      },
      "TMDBGenreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "MovieColorsDto": {
        "type": "object",
        "properties": {
          "primary": {
            "type": "string"
          },
          "secondary": {
            "type": "string"
          },
          "accent": {
            "type": "string"
          },
          "muted": {
            "type": "string"
          }
        }
      },
      "TMDBCastDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "character": {
            "type": "string"
          },
          "profile_path": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "order"
        ]
      },
...
      "TMDBCreditsDto": {
        "type": "object",
        "properties": {
          "cast": {
            "description": "Top-billed cast only; see cast_total and the credits endpoint",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TMDBCastDto"
            }
          },
          "crew": {
            "description": "Key crew only; see crew_total and the credits endpoint",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TMDBCrewDto"
            }
          },
          "cast_total": {
            "type": "number",
            "description": "Cast members TMDB lists in total"
          },
          "crew_total": {
            "type": "number",
            "description": "Crew credits TMDB lists in total"
          }
        },
        "required": [
          "cast",
          "crew",
          "cast_total",
          "crew_total"
        ]
      },
      "TMDBTrailerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "site": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "official": {
            "type": "boolean"
          },
          "published_at": {
            "type": "string"
          },
          "sourceMediaType": {
            "type": "string",
            "enum": [
              "movie",
              "show",
              "season",
              "episode"
            ]
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "site",
          "type",
          "sourceMediaType"
        ]
      },
      "TMDBMovieDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "poster_path": {
            "type": "string"
          },
          "backdrop_path": {
            "type": "string"
          },
          "release_date": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "vote_average": {
            "type": "number"
          },
          "runtime": {
            "type": "number"
          },
          "vote_count": {
            "type": "number"
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TMDBGenreDto"
            }
          },
          "colors": {
            "$ref": "#/components/schemas/MovieColorsDto"
          },
          "credits": {
            "$ref": "#/components/schemas/TMDBCreditsDto"
          },
          "trailer": {
            "$ref": "#/components/schemas/TMDBTrailerDto"
          }
        },
        "required": [
          "id",
          "title"
        ]
      },
      "CrewDepartmentDto": {
        "type": "object",
        "properties": {
          "department": {
            "type": "string",
            "description": "Display label, e.g. \"Directing\", \"Other crew\""
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TMDBCrewDto"
            }
          }
        },
        "required": [
          "department",
          "members"
        ]
      },
      "FullCreditsDto": {
        "type": "object",
        "properties": {
          "cast": {
            "description": "Every cast member, by billing",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TMDBCastDto"
            }
          },
          "crew": {
            "description": "Every crew credit, by department",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CrewDepartmentDto"
            }
          }
        },
        "required": [
          "cast",
          "crew"
        ]
      },
...
      "WatchProvidersResultDto": {
        "type": "object",
        "properties": {
          "flatrate": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatchProviderDto"
            }
          },
          "rent": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatchProviderDto"
            }
          },
          "buy": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatchProviderDto"
            }
          },
          "ads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatchProviderDto"
            }
          },
          "free": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatchProviderDto"
            }
          }
        }
      },
      "WatchProvidersResponseDto": {
        "type": "object",
        "properties": {
          "providers": {
            "$ref": "#/components/schemas/WatchProvidersResultDto"
          },
          "availableCountries": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MovieDto": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "posterPath": {
            "type": "string"
          },
          "backdropPath": {
            "type": "string"
          },
          "releaseYear": {
            "type": "number"
          },
          "releaseDate": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "colors": {
            "$ref": "#/components/schemas/MovieColorsDto"
          }
        },
        "required": [
          "movieId",
          "title"
        ]
      },
      "TrackedMovieDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "rkey": {
            "type": "string"
          },
          "uri": {
            "type": "string"
          },
          "cid": {
            "type": "string"
          },
          "userDid": {
            "type": "string"
          },
          "movieId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "watchedDate": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "movie": {
            "$ref": "#/components/schemas/MovieDto"
          }
        },
        "required": [
          "id",
          "rkey",
          "uri",
          "cid",
          "userDid",
          "movieId",
          "status",
          "createdAt",
          "updatedAt",
          "movie"
        ]
      },
      "WatchHistoryItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "watchedDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "UnifiedSearchResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "media_type": {
            "type": "string",
            "enum": [
              "movie",
              "tv"
            ]
          },
          "title": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "poster_path": {
            "type": "string"
          },
          "backdrop_path": {
            "type": "string"
          },
          "release_date": {
            "type": "string"
          },
          "first_air_date": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "popularity": {
            "type": "number"
          },
          "vote_average": {
            "type": "number"
          },
          "vote_count": {
            "type": "number"
          },
          "original_language": {
            "type": "string"
          },
          "genre_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "original_title": {
            "type": "string"
          },
          "original_name": {
            "type": "string"
          },
          "adult": {
            "type": "boolean"
          },
          "video": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "media_type",
          "popularity",
          "vote_average",
          "vote_count"
        ]
      },
...
      "UnifiedDiscoverResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total count of items across all pages"
          },
          "page": {
            "type": "number",
            "description": "Current page number after server-side clamping"
          },
          "pageSize": {
            "type": "number",
            "description": "Number of items per page"
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of available pages"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Whether a next page exists"
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "Whether a previous page exists"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnifiedSearchResultDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage",
          "items"
        ]
      },
...
    }
  }
}